+++ ui/components/filters/custom-provider-inputs.tsx
@@ -5,6 +5,7 @@ import {
   AzureProviderBadge,
   GCPProviderBadge,
   GitHubProviderBadge,
+  IacProviderBadge,
   KS8ProviderBadge,
   M365ProviderBadge,
 } from "../icons/providers-badge";
@@ -59,6 +60,15 @@ export const CustomProviderInputGitHub = () => {
     <div className="flex items-center gap-x-2">
       <GitHubProviderBadge width={25} height={25} />
       <p className="text-sm">GitHub</p>
+    </div>
+  );
+};
+
+export const CustomProviderInputIac = () => {
+  return (
+    <div className="flex items-center gap-x-2">
+      <IacProviderBadge width={25} height={25} />
+      <p className="text-sm">Infrastructure as Code</p>
     </div>
   );
 };