+++ ui/components/providers/workflow/forms/via-credentials/index.ts
@@ -1,4 +1,5 @@
 export * from "./azure-credentials-form";
 export * from "./github-credentials-form";
+export * from "./iac-credentials-form";
 export * from "./k8s-credentials-form";
 export * from "./m365-credentials-form";